GC refactor - Moved garbageCollection.ts and associated files into a sub-directory
