+++ packages/runtime/container-runtime/src/gc/index.ts
@@ -1,0 +1,14 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export {
+    gcBlobPrefix,
+    GCNodeType,
+    gcTreeKey,
+    IGarbageCollectionRuntime,
+    IGarbageCollector,
+    IGCStats,
+} from "./gcDefinitions";
+export { GarbageCollector } from "./garbageCollection";
