+++ packages/runtime/container-runtime/src/test/gcUnreferencedStateTracker.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-internal-modules */
 /*!
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
@@ -7,10 +8,8 @@
 
 import { strict as assert } from "assert";
 import { SinonFakeTimers, SinonSpy, useFakeTimers, spy } from "sinon";
-import {
-    UnreferencedState,
-    UnreferencedStateTracker,
-} from "../garbageCollection";
+import { UnreferencedState } from "../gc/gcDefinitions";
+import { UnreferencedStateTracker } from "../gc/gcUnreferencedStateTracker";
 
 describe("Garbage Collection Tests", () => {
     let clock: SinonFakeTimers;
